feat: implementando a exceção DELETE

+++ ms_event_manager/src/main/java/com/compass/ms_event_manager/controller/EventController.java
@@ -6,8 +6,10 @@ import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/events/v1")
@@ -15,6 +17,9 @@ public class EventController {
 
     @Autowired
     private EventService service;
+
+    @Autowired
+    private RestTemplate restTemplate;
 
     @PostMapping("/create-event")
     public ResponseEntity<Event> createEvent(@Valid @RequestBody Event event) {
@@ -39,9 +44,19 @@ public class EventController {
     }
 
     @DeleteMapping("/delete-event/{id}")
-    public ResponseEntity<Void> deleteEvent(@PathVariable String id) {
+    public ResponseEntity<?> deleteEvent(@PathVariable String id) {
+        String url = "http://localhost:8081/tickets/v1/check-tickets-by-event/" + id;
+        Boolean hasTickets = restTemplate.getForObject(url, Boolean.class);
+
+        if (Boolean.TRUE.equals(hasTickets)) {
+            return ResponseEntity.status(409)
+                    .body(Map.of("error", "O evento não pode ser deletado porque possui ingressos vendidos."));
+        }
+
         service.deleteEvent(id);
-        return ResponseEntity.noContent().build();
+        return ResponseEntity.ok(Map.of("message", "Evento deletado com sucesso!"));
     }
+
+
 }
